+++ cel/migrated/branches/R0_98/plugins/behaviourlayer/xml/blxml.cpp
@@ -796,7 +796,7 @@ bool celBlXml::ParseExpressionInt (
 	}
 	size_t i;
 	size_t varidx = (size_t)-1;
-	for (i = 0 ; i < local_vars.Length () ; i++)
+	for (i = 0 ; i < (size_t)local_vars.Length () ; i++)
 	  if (!strcmp (str, local_vars[i]))
 	  {
 	    varidx = i;
@@ -1617,7 +1617,7 @@ bool celBlXml::ParseEventHandler (celXmlScriptEventHandler* h,
 	  }
 	  size_t varidx = (size_t)-1;
 	  size_t i;
-	  for (i = 0 ; i < local_vars.Length () ; i++)
+	  for (i = 0 ; i < (size_t)local_vars.Length () ; i++)
 	    if (!strcmp (varname, local_vars[i]))
 	    {
 	      varidx = i;
@@ -1756,7 +1756,7 @@ bool celBlXml::ParseEventHandler (celXmlScriptEventHandler* h,
 	  }
 	  size_t end_idx = h->GetLastArgumentIndex ()+1;
 	  size_t i;
-	  for (i = 0 ; i < goto_end_idx.Length () ; i++)
+	  for (i = 0 ; i < (size_t)goto_end_idx.Length () ; i++)
 	  {
 	    h->GetArgument (goto_end_idx[i]).SetCodeLocation (end_idx);
 	  }
