+++ cel/migrated/branches/R0_98/plugins/propclass/tools/toolfact.cpp
@@ -368,7 +368,7 @@ celPcProperties::~celPcProperties ()
 size_t celPcProperties::FindProperty (csStringID id)
 {
   size_t i;
-  for (i = 0 ; i < properties.Length () ; i++)
+  for (i = 0 ; i < (size_t)properties.Length () ; i++)
   {
     property* p = properties[i];
     if (p->id == csInvalidStringID)
@@ -548,7 +548,7 @@ csPtr<iCelDataBuffer> celPcProperties::Save ()
   csRef<iCelDataBuffer> databuf = pl->CreateDataBuffer (PROPERTIES_SERIAL);
   databuf->SetDataCount (properties.Length ()*3);
   size_t i, j = 0;
-  for (i = 0 ; i < properties.Length () ; i++)
+  for (i = 0 ; i < (size_t)properties.Length () ; i++)
   {
     property* p = properties[i];
     databuf->GetData (j++)->Set (p->propName);
@@ -730,7 +730,7 @@ size_t celPcProperties::GetPropertyIndex (const char* name)
     // Create the hash.
     properties_hash.DeleteAll ();
     size_t i;
-    for (i = 0 ; i < properties.Length () ; i++)
+    for (i = 0 ; i < (size_t)properties.Length () ; i++)
     {
       property* p = properties[i];
       properties_hash.Put (p->propName, i+1);	// +1 so that 0 is an error
@@ -742,7 +742,7 @@ size_t celPcProperties::GetPropertyIndex (const char* name)
 
 void celPcProperties::SetPropertyIndex (size_t index, float value)
 {
-  CS_ASSERT ((index != csArrayItemNotFound) && index < properties.Length ());
+  CS_ASSERT ((index != csArrayItemNotFound) && index < (size_t)properties.Length ());
   property* p = properties[index];
   ClearPropertyValue (p);
   p->type = CEL_DATA_FLOAT;
@@ -758,7 +758,7 @@ void celPcProperties::SetPropertyIndex (size_t index, float value)
 
 void celPcProperties::SetPropertyIndex (size_t index, long value)
 {
-  CS_ASSERT (index >= 0 && index < properties.Length ());
+  CS_ASSERT (index >= 0 && index < (size_t)properties.Length ());
   property* p = properties[index];
   ClearPropertyValue (p);
   p->type = CEL_DATA_LONG;
@@ -774,7 +774,7 @@ void celPcProperties::SetPropertyIndex (size_t index, long value)
 
 void celPcProperties::SetPropertyIndex (size_t index, bool value)
 {
-  CS_ASSERT (index >= 0 && index < properties.Length ());
+  CS_ASSERT (index >= 0 && index < (size_t)properties.Length ());
   property* p = properties[index];
   ClearPropertyValue (p);
   p->type = CEL_DATA_BOOL;
@@ -790,7 +790,7 @@ void celPcProperties::SetPropertyIndex (size_t index, bool value)
 
 void celPcProperties::SetPropertyIndex (size_t index, const csVector2& value)
 {
-  CS_ASSERT (index >= 0 && index < properties.Length ());
+  CS_ASSERT (index >= 0 && index < (size_t)properties.Length ());
   property* p = properties[index];
   ClearPropertyValue (p);
   p->type = CEL_DATA_VECTOR2;
@@ -807,7 +807,7 @@ void celPcProperties::SetPropertyIndex (size_t index, const csVector2& value)
 
 void celPcProperties::SetPropertyIndex (size_t index, const csVector3& value)
 {
-  CS_ASSERT (index >= 0 && index < properties.Length ());
+  CS_ASSERT (index >= 0 && index < (size_t)properties.Length ());
   property* p = properties[index];
   ClearPropertyValue (p);
   p->type = CEL_DATA_VECTOR3;
@@ -825,7 +825,7 @@ void celPcProperties::SetPropertyIndex (size_t index, const csVector3& value)
 
 void celPcProperties::SetPropertyIndex (size_t index, const csColor& value)
 {
-  CS_ASSERT (index >= 0 && index < properties.Length ());
+  CS_ASSERT (index >= 0 && index < (size_t)properties.Length ());
   property* p = properties[index];
   ClearPropertyValue (p);
   p->type = CEL_DATA_COLOR;
@@ -843,7 +843,7 @@ void celPcProperties::SetPropertyIndex (size_t index, const csColor& value)
 
 void celPcProperties::SetPropertyIndex (size_t index, const char* value)
 {
-  CS_ASSERT (index >= 0 && index < properties.Length ());
+  CS_ASSERT (index >= 0 && index < (size_t)properties.Length ());
   property* p = properties[index];
   ClearPropertyValue (p);
   p->type = CEL_DATA_STRING;
@@ -859,7 +859,7 @@ void celPcProperties::SetPropertyIndex (size_t index, const char* value)
 
 void celPcProperties::SetPropertyIndex (size_t index, iCelPropertyClass* value)
 {
-  CS_ASSERT (index >= 0 && index < properties.Length ());
+  CS_ASSERT (index >= 0 && index < (size_t)properties.Length ());
   property* p = properties[index];
   ClearPropertyValue (p);
   p->type = CEL_DATA_PCLASS;
@@ -875,7 +875,7 @@ void celPcProperties::SetPropertyIndex (size_t index, iCelPropertyClass* value)
 
 void celPcProperties::SetPropertyIndex (size_t index, iCelEntity* value)
 {
-  CS_ASSERT (index >= 0 && index < properties.Length ());
+  CS_ASSERT (index >= 0 && index < (size_t)properties.Length ());
   property* p = properties[index];
   ClearPropertyValue (p);
   p->type = CEL_DATA_ENTITY;
@@ -891,14 +891,14 @@ void celPcProperties::SetPropertyIndex (size_t index, iCelEntity* value)
 
 celDataType celPcProperties::GetPropertyType (size_t index) const
 {
-  CS_ASSERT (index >= 0 && index < properties.Length ());
+  CS_ASSERT (index >= 0 && index < (size_t)properties.Length ());
   property* p = properties[index];
   return p->type;
 }
 
 float celPcProperties::GetPropertyFloatIndex (size_t index) const
 {
-  CS_ASSERT (index >= 0 && index < properties.Length ());
+  CS_ASSERT (index >= 0 && index < (size_t)properties.Length ());
   property* p = properties[index];
   if (p->type == CEL_DATA_FLOAT)
     return p->v.f;
@@ -908,7 +908,7 @@ float celPcProperties::GetPropertyFloatIndex (size_t index) const
 
 long celPcProperties::GetPropertyLongIndex (size_t index) const
 {
-  CS_ASSERT (index >= 0 && index < properties.Length ());
+  CS_ASSERT (index >= 0 && index < (size_t)properties.Length ());
   property* p = properties[index];
   if (p->type == CEL_DATA_LONG)
     return p->v.l;
@@ -918,7 +918,7 @@ long celPcProperties::GetPropertyLongIndex (size_t index) const
 
 bool celPcProperties::GetPropertyBoolIndex (size_t index) const
 {
-  CS_ASSERT (index >= 0 && index < properties.Length ());
+  CS_ASSERT (index >= 0 && index < (size_t)properties.Length ());
   property* p = properties[index];
   if (p->type == CEL_DATA_BOOL)
     return p->v.b;
@@ -928,7 +928,7 @@ bool celPcProperties::GetPropertyBoolIndex (size_t index) const
 
 bool celPcProperties::GetPropertyVectorIndex (size_t index, csVector2& v) const
 {
-  CS_ASSERT (index >= 0 && index < properties.Length ());
+  CS_ASSERT (index >= 0 && index < (size_t)properties.Length ());
   property* p = properties[index];
   if (p->type == CEL_DATA_VECTOR2)
   {
@@ -942,7 +942,7 @@ bool celPcProperties::GetPropertyVectorIndex (size_t index, csVector2& v) const
 
 bool celPcProperties::GetPropertyVectorIndex (size_t index, csVector3& v) const
 {
-  CS_ASSERT (index >= 0 && index < properties.Length ());
+  CS_ASSERT (index >= 0 && index < (size_t)properties.Length ());
   property* p = properties[index];
   if (p->type == CEL_DATA_VECTOR3)
   {
@@ -957,7 +957,7 @@ bool celPcProperties::GetPropertyVectorIndex (size_t index, csVector3& v) const
 
 bool celPcProperties::GetPropertyColorIndex (size_t index, csColor& v) const
 {
-  CS_ASSERT (index >= 0 && index < properties.Length ());
+  CS_ASSERT (index >= 0 && index < (size_t)properties.Length ());
   property* p = properties[index];
   if (p->type == CEL_DATA_COLOR)
   {
@@ -972,7 +972,7 @@ bool celPcProperties::GetPropertyColorIndex (size_t index, csColor& v) const
 
 const char* celPcProperties::GetPropertyStringIndex (size_t index) const
 {
-  CS_ASSERT (index >= 0 && index < properties.Length ());
+  CS_ASSERT (index >= 0 && index < (size_t)properties.Length ());
   property* p = properties[index];
   if (p->type == CEL_DATA_STRING)
     return p->v.s;
@@ -982,7 +982,7 @@ const char* celPcProperties::GetPropertyStringIndex (size_t index) const
 
 iCelPropertyClass* celPcProperties::GetPropertyPClassIndex (size_t index) const
 {
-  CS_ASSERT (index >= 0 && index < properties.Length ());
+  CS_ASSERT (index >= 0 && index < (size_t)properties.Length ());
   property* p = properties[index];
   if (p->type == CEL_DATA_PCLASS)
     return p->pclass;
@@ -992,7 +992,7 @@ iCelPropertyClass* celPcProperties::GetPropertyPClassIndex (size_t index) const
 
 iCelEntity* celPcProperties::GetPropertyEntityIndex (size_t index) const
 {
-  CS_ASSERT (index >= 0 && index < properties.Length ());
+  CS_ASSERT (index >= 0 && index < (size_t)properties.Length ());
   property* p = properties[index];
   if (p->type == CEL_DATA_ENTITY)
     return p->entity;
@@ -1002,7 +1002,7 @@ iCelEntity* celPcProperties::GetPropertyEntityIndex (size_t index) const
 
 void celPcProperties::ClearProperty (size_t index)
 {
-  CS_ASSERT (index >= 0 && index < properties.Length ());
+  CS_ASSERT (index >= 0 && index < (size_t)properties.Length ());
   iCelBehaviour* bh = entity->GetBehaviour ();
   if (bh)
   {
@@ -1034,7 +1034,7 @@ size_t celPcProperties::GetPropertyCount () const
 
 const char* celPcProperties::GetPropertyName (size_t index) const
 {
-  CS_ASSERT (index >= 0 && index < properties.Length ());
+  CS_ASSERT (index >= 0 && index < (size_t)properties.Length ());
   property* p = properties[index];
   return p->propName;
 }
